safari-534.53-branch/LayoutTests/storage/sql-error-codes-expected.txt

CONSOLE MESSAGE: line 0: Exception thrown in transaction callback.
CONSOLE MESSAGE: line 0: Cannot call toString() on this object.
This test tests the error codes reported in exceptional situations.
PASS: expected and got error code UNKNOWN_ERR
PASS: expected and got error code UNKNOWN_ERR
PASS: expected and got error code SYNTAX_ERR
PASS: expected and got error code SYNTAX_ERR
PASS: expected and got error code UNKNOWN_ERR
PASS: expected and got error code QUOTA_ERR
PASS: expected and got error code VERSION_ERR
